Move process for rete functions out of export.ts and into dataflow-program-model.ts
