+++ src/plugins/dataflow-tool/model/utilities/export.ts
@@ -1,7 +1,4 @@
-import { cloneDeep } from "lodash";
 import { DataSequence } from "../../components/ui/dataflow-program-graph";
-import { DataflowProgramSnapshotOut, DataflowNodeSnapshotOut, DataflowSocketSnapshotOut }
-  from "../dataflow-program-model";
 
 export const exportDataCSV = (dataSequences: DataSequence[]) => {
   if (!dataSequences) {
@@ -38,49 +35,4 @@ export const exportCSV = (csv: string, fileName: string) => {
   // link created by us in code, so calling click on the link should not trigger blockers
   link.click();
   document.body.removeChild(link);
-};
-
-// The following functions are used to convert a program snapshot from MST/firebase to
-// a format rete will accept. See comments for preProcessSnapshot() functions in
-// dataflow-program-model.ts to see what these functions are undoing.
-
-const postProcessSocketSnapshotForRete = (snapshot: DataflowSocketSnapshotOut) => {
-  return { connections: Object.values(snapshot.connections) };
-};
-
-const postProcessSocketsSnapshotForRete = (snapshot: Record<string, DataflowSocketSnapshotOut>) => {
-  const processedSockets: any = {};
-  for (const key in snapshot) {
-    processedSockets[key] = postProcessSocketSnapshotForRete(snapshot[key]);
-  }
-  return processedSockets;
-};
-
-const postProcessNodeSnapshotForRete = (snapshot: DataflowNodeSnapshotOut) => {
-  const { x, y, inputs, outputs, ...rest } = snapshot;
-  return {
-    position: [x, y],
-    inputs: postProcessSocketsSnapshotForRete(inputs),
-    outputs: postProcessSocketsSnapshotForRete(outputs),
-    ...rest
-  };
-};
-
-export const postProcessProgramSnapshotForRete = (snapshot: DataflowProgramSnapshotOut) => {
-  const { nodes, values, ...rest } = snapshot;
-  const newNodes = cloneDeep(nodes) as any;
-  const keys = Object.keys(newNodes);
-  keys.forEach((key: string) => {
-    newNodes[key] = postProcessNodeSnapshotForRete(newNodes[key]);
-    const data = newNodes[key].data;
-    if ((values as any)?.[key]) {
-      const { nodeValue, recentValues } = (values as any)[key];
-      data.nodeValue = nodeValue;
-      data.recentValues = JSON.parse(recentValues);
-    } else {
-      data.nodeValue = null;
-      data.recentValues = [];
-    }
-  });
-  return { nodes: newNodes, ...rest };
 };
